% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{AnalysisData}
\alias{AnalysisData}
\title{Create a new AnalysisData object}
\usage{
AnalysisData(filepath, ...)
}
\arguments{
\item{filepath}{Path to the CSV data file.}

\item{...}{Additional arguments passed to read.csv.}
}
\value{
A new AnalysisData object with data loaded from the CSV file.
}
\description{
Create a new AnalysisData object
}
