% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{calculateFoldChange}
\alias{calculateFoldChange}
\title{Calculate Fold Change}
\usage{
calculateFoldChange(object, idColumn, vehicleSamples, treatmentSamples, ...)
}
\arguments{
\item{object}{A ProteinAnalysisData object.}

\item{idColumn}{The Protein ID column name (character).}

\item{vehicleSamples}{The columns containing the normalized vehicle intensities (list).}

\item{treatmentSamples}{The columns containing the normalize treatment intensities (list).}
}
\value{
The updated ProteinAnalysisData object.
}
\description{
Calculate Fold Change
}
