% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProteinDataAnalysis.R
\name{cleanData}
\alias{cleanData}
\title{Remove unwanted ProteinIDs in ProteinAnalysisData}
\usage{
cleanData(object, idColumn, cleanUpIDs, ...)
}
\arguments{
\item{object}{A ProteinAnalysisData object.}

\item{idColumn}{The Protein ID column name (character) to normalize.}

\item{cleanUpIDs}{The Protein IDs to be removed (list).}
}
\value{
The updated ProteinAnalysisData object.
}
\description{
Remove unwanted ProteinIDs in ProteinAnalysisData
}
