% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProteinDataAnalysis.R
\docType{data}
\name{commonContaminantProteinIDs}
\alias{commonContaminantProteinIDs}
\title{Standard set of ProteinIDs to remove (e.g., contaminants or unwanted proteins)}
\format{
An object of class \code{character} of length 1.
}
\usage{
commonContaminantProteinIDs
}
\description{
Standard set of ProteinIDs to remove (e.g., contaminants or unwanted proteins)
}
\keyword{datasets}
