% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProteinDataAnalysis.R
\name{getDataframe}
\alias{getDataframe}
\title{Get dataframe}
\usage{
getDataframe(object, ...)
}
\arguments{
\item{object}{A ProteinAnalysisData object.}
}
\value{
Dataframe of object data
}
\description{
Get dataframe
}
