% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{read.fragpipe}
\alias{read.fragpipe}
\title{Create a new ProteinAnalysisData object}
\usage{
read.fragpipe(filepath, ...)
}
\arguments{
\item{filepath}{Path to the CSV data file.}

\item{...}{Additional arguments passed to read.csv.}
}
\value{
A new ProteinAnalysisData object with data loaded from the CSV file.
}
\description{
Create a new ProteinAnalysisData object
}
