% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProteinDataAnalysis.R
\name{removeContaminants}
\alias{removeContaminants}
\title{Remove unwanted ProteinIDs in ProteinAnalysisData}
\usage{
removeContaminants(object, idColumn, cleanUpIDs, ...)
}
\arguments{
\item{object}{A ProteinAnalysisData object.}

\item{idColumn}{Character. The column name containing Protein IDs.}

\item{cleanUpIDs}{Character vector of Protein IDs to remove.}
}
\value{
The updated ProteinAnalysisData object.
}
\description{
Removes rows where the value in idColumn matches any of the provided cleanUpIDs.
}
