% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProteinDataAnalysis.R
\name{runCorrectedTTest}
\alias{runCorrectedTTest}
\title{Run t-test}
\usage{
runCorrectedTTest(
  object,
  vehicleSamples,
  treatmentSamples,
  method = "Benjamini-Hochberg",
  ...
)
}
\arguments{
\item{object}{A ProteinAnalysisData object.}

\item{vehicleSamples}{The columns containing the normalized vehicle intensities (list).}

\item{treatmentSamples}{The columns containing the normalize treatment intensities (list).}
}
\value{
The updated ProteinAnalysisData object.
}
\description{
Run t-test
}
