% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{runtTest}
\alias{runtTest}
\title{Run t-test}
\usage{
runtTest(object, vehicleSamples, treatmentSamples, ...)
}
\arguments{
\item{object}{A ProteinAnalysisData object.}

\item{vehicleSamples}{The columns containing the normalized vehicle intensities (list).}

\item{treatmentSamples}{The columns containing the normalize treatment intensities (list).}
}
\value{
The updated ProteinAnalysisData object.
}
\description{
Run t-test
}
