% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProteinDataAnalysis.R
\name{setSamples}
\alias{setSamples}
\title{Set Samples}
\usage{
setSamples(object, vehicleSamples, treatmentSamples, ...)
}
\arguments{
\item{object}{A ProteinAnalysisData object.}

\item{vehicleSamples}{Character vector of column names for vehicle samples.}

\item{treatmentSamples}{Character vector of column names for treatment samples.}
}
\value{
The updated ProteinAnalysisData object.
}
\description{
Set Samples
}
