% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProteinDataAnalysis.R
\name{volcanoPlot}
\alias{volcanoPlot}
\title{Generate Volcano Plot}
\usage{
volcanoPlot(
  object,
  idColumn,
  foldChange = "FoldChange",
  pValueCol = "TTestPValue",
  ...
)
}
\arguments{
\item{object}{A ProteinAnalysisData object.}

\item{idColumn}{Column containing ProteinIDs (character scalar)}

\item{foldChange}{Column containing fold change values (character scalar)}
}
\value{
A ggplot2 object
}
\description{
Generate Volcano Plot
}
