% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProteinDataAnalysis.R
\name{baitNormalize}
\alias{baitNormalize}
\title{Normalize to Bait}
\usage{
baitNormalize(object, idColumn, cleanUpIDs, removeBait = TRUE, ...)
}
\arguments{
\item{object}{A ProteinAnalysisData object.}

\item{idColumn}{The Protein ID column name (character) to normalize.}

\item{removeBait}{Remove the Bait entry from the object (boolean).}

\item{baitID}{The UniProt Protein ID of the Bait used (character).}
}
\value{
The updated ProteinAnalysisData object.
}
\description{
Normalize to Bait
}
