% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProteinDataAnalysis.R
\name{calculateFoldChange}
\alias{calculateFoldChange}
\title{Calculate Fold Change}
\usage{
calculateFoldChange(
  object,
  idColumn,
  vehicleSamples = NULL,
  treatmentSamples = NULL,
  ...
)
}
\arguments{
\item{object}{A ProteinAnalysisData object.}

\item{idColumn}{The Protein ID column name (character).}

\item{vehicleSamples}{The columns containing the normalized vehicle intensities (character vector; optional).}

\item{treatmentSamples}{The columns containing the normalized treatment intensities (character vector; optional).}
}
\value{
The updated ProteinAnalysisData object.
}
\description{
Calculate Fold Change
}
