% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProteinDataAnalysis.R
\name{runCorrectedTTest}
\alias{runCorrectedTTest}
\title{Run t-test}
\usage{
runCorrectedTTest(
  object,
  vehicleSamples = NULL,
  treatmentSamples = NULL,
  method = "Benjamini-Hochberg",
  ...
)
}
\arguments{
\item{object}{A ProteinAnalysisData object.}

\item{vehicleSamples}{The columns containing the normalized vehicle intensities (character vector; optional).}

\item{treatmentSamples}{The columns containing the normalized treatment intensities (character vector; optional).}

\item{method}{Correction method for multiple comparison (default: "Benjamini-Hochberg").}
}
\value{
The updated ProteinAnalysisData object.
}
\description{
Run t-test
}
