% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProteinDataAnalysis.R
\name{volcanoPlot}
\alias{volcanoPlot}
\title{Generate Volcano Plots}
\usage{
volcanoPlot(object, idColumn, foldChange, ...)
}
\arguments{
\item{object}{A ProteinAnalysisData object.}

\item{idColumn}{Column containing ProteinIDs}

\item{foldChange}{Column containing fold change values}
}
\value{
A ggplot2 object
}
\description{
Generate Volcano Plots
}
